Declare a global count variable
I wanted to fetch the start classification count and then add to it
incremntally. As I have moved everything into async functions, I have
started a global count var so that this can be accessed at all times to
be added to and written to the appropriate elements on the page.

This is a work in progress but at this point the project&user count (in
the centre of the page) works.