Get project and user ID from url query
As opposed to manually putting in the desired project and user to track in the script, instead pull these numbers from a url query. This will make the tally flexible to be easily and quickly used for different projects.

To do this I used on window.location.search and URLSearchParams.