Add a function to query by project name
A function has been added for querying by username. It follows the same
logic as convertUsernameToID(). The crux of this is that if no project
ID is given in the URL but there is a project name then the ID is looked
up and overwrites the global varibale projectID. The new projectID
variable is however local to the function scope therefore it must be
passed to all subsequent functions, else the global projectID will be
used (which is null as no project_id was supplied in the URL query).

While the logic is there, this is not currently functional due to the
Panoptes API query by project name being case sensitive. The project ID
supplied in the query URL is by default converted to lowercase.

Ideally users of the App will search by project name lower case and
hyphenated (etch-a-cell). In the function the spaces in this string are
converted to spaces to match the API, however a sensible way of
implementing the correct case has not been found. For instance the API
query for this project would need to be 'Etch a Cell'.